Update rand package version dependency.
Prior to this commit attempting to build the project with `cargo build
--release` failed to compile the rand dependency (which was defaulting to v0.3.8).
After this commit the rand dependency requires a more current version
and the project builds successfully.

I'm working on learning some Rust myself and wanted to turn some knobs
in your project, this fix allowed me to build it.